move models to common package
