+++ internal/core/models/location.go
@@ -1,0 +1,6 @@
+package models
+
+type LocationResponse struct {
+	PostalCode  string `json:"postalCode"`
+	CountryCode string `json:"countryCode"`
+}
