--- internal/core/models/offer_request.go
@@ -1,5 +1,0 @@
-package models
-
-type OfferRequest struct {
-	UserDeviceID string `json:"user_device_id"`
-}
